fix: naming and private filterMessage
