+++ pkg/message_test.go
@@ -2,27 +2,8 @@ package pkg
 
 import (
 	"os"
-	"regexp"
 	"testing"
 )
-
-func TestFilterMessage(t *testing.T) {
-	skips := []*regexp.Regexp{
-		regexp.MustCompile("shouldmatch"),
-		regexp.MustCompile("wontmatch"),
-	}
-	tests := map[string]bool{
-		"nomatch":                               true,
-		"i shouldmatch, because i should match": false,
-		"i won't actually match":                true,
-	}
-
-	for input, expected := range tests {
-		if FilterMessage(input, skips) != expected {
-			t.Fatalf("expected %t result for: %s", expected, input)
-		}
-	}
-}
 
 func TestMessageRenderPrefaceNoSubject(t *testing.T) {
 	msg := &Message{
